backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
pymysql==1.1.1
psycopg2-binary==2.9.10
pyodbc==5.2.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
itsdangerous==2.2.0
resend==2.2.0
jinja2==3.1.4
bcrypt==4.0.1
